\name{TruncComp-package}
\alias{TruncComp-package}
\alias{TruncComp}
\docType{package}

\title{\packageTitle{TruncComp}}

\description{\packageDescription{TruncComp}}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{TruncComp}
\packageIndices{TruncComp}
~~ An overview of how to use the package, including the most important functions ~~
}

\author{
\packageAuthor{TruncComp}

Maintainer: \packageMaintainer{TruncComp}
}


\references{
~~ Literature or other references for background information ~~
}


\keyword{ package }

\examples{
1 + 2
}
